Add Abseil (#32)
This pull request introduces a minor update to the project's
dependencies. The most significant change is the addition of the
`abseil` library to the general dependencies in `vcpkg.json`.

- Added `abseil` as a general dependency in `vcpkg.json` to support
future development needs.
This pull request makes a small change to the project's dependencies by
adding the `abseil` library to the general dependencies in `vcpkg.json`.